Refactor endpoint network runtime helpers
